requirements.txt
speedtest_cli==2.1.2
influxdb-client==1.12.0